move test for Enumerator::Lazy#compact to compact_spec file
